requirements.txt
langchain
langchain-nvidia-ai-endpoints
faiss-cpu
langchain-community
streamlit
arxiv